static_cast to remove warnings; JIRA: PGAP-9240
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@102886 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/algo/align/nw/nw_spliced_aligner32.cpp
@@ -389,7 +389,7 @@ CNWAligner::TScore CSplicedAligner32::x_Align (SAlignInOut* data)
                 type = kTypeIntron | dnr_pos;
             }
 
-            backtrace_matrix[k] = type;
+            backtrace_matrix[k] = static_cast<unsigned int>(type);
 
             // detect donor candidates
             if(j < N2 - 2) {
